+++ CHANGELOG.rst
@@ -1,5 +1,5 @@
-Unreleased
-----------
+v0.45.2
+-------
 - Detect unclosed comments, e.g. ``<!== a``.
 - Fix parsing priority of tag extensions and comments. For example the comment in ``<ref>b<!--c</ref>d-->`` used to be parsed as with ``<!--c</ref>d-->`` as comment which was incorrect.